Symlink all the third-party libraries to the top-level.
We do this instead of modifying sys.path because AppEngine infrastructure -- in
particular deferred tasks -- needs to be able to be able to find all our imports
without running pub_dartlang.py.

This fixes a bug where deferred tasks would fail due to being unable to find
cherrypy. This was causing some reload failures, and probably #74 as well.