Get all of the tests working again.
When we upgraded to AppEngine 1.7.6, we changed the way
in_production() was detected. That tickled the code in the tests
that faked the production environment so that the error page tests
would not think they were running on a dev server.

This fixes that. It also fixes a couple of other tests that were
broken. In cases where there isn't a valid route (i.e. a completely
unrecognized URL), Request._package_name would raise an error. Now
it returns None as it should.

With this patch, all tests are passing on my machine. Yay!